agent/requirements.txt
langchain==0.1.10
langchain-openai==0.0.8
langgraph==0.0.25
web3==6.11.3
eth-account==0.10.0
python-telegram-bot==20.5
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
